added functionality of editing the photo
